Update migration tables.

+++ database/migrations/2023_08_26_135027_add_categories_fk_to_products.php
@@ -11,12 +11,15 @@ return new class extends Migration
      */
     public function up(): void
     {
-        Schema::table('products', function (Blueprint $table) {
+        if (!Schema::hasTable('products')) {
 
-            $table->unsignedBigInteger('category_id');
+            Schema::table('products', function (Blueprint $table) {
 
-            $table->foreign('category_id')->references('id')->on('categories')->onDelete('cascade');
-        });
+                $table->unsignedBigInteger('category_id');
+
+                $table->foreign('category_id')->references('id')->on('categories')->onDelete('cascade');
+            });
+        }
     }
 
     /**
